requirements.txt
Django>=5.1,<5.2
django-environ
django-formtools
django-qrcode
django-widget-tweaks
google-generativeai
groq
langchain
langchain-groq
langchain-text-splitters
faiss-cpu
sentence-transformers
torch
transformers
pdfkit
pdfplumber
PyPDF2
pillow
qrcode
markdown
markdown2
pandas
numpy
scikit-learn
scipy
pydantic
